+++ ci/cpu/build.sh
@@ -16,6 +16,13 @@ export MINOR_VERSION=`echo $GIT_DESCRIBE | grep -o -E '([0-9]+\.[0-9]+)'`
 # Set home to the job's workspace
 export HOME=$WORKSPACE
 
+# Determine CUDA release version
+export CUDA_REL=${CUDA_VERSION%.*}
+
+# Setup 'gpuci_conda_retry' for build retries (results in 2 total attempts)
+export GPUCI_CONDA_RETRY_MAX=1
+export GPUCI_CONDA_RETRY_SLEEP=30
+
 # Switch to project root; also root of repo checkout
 cd $WORKSPACE
 
@@ -32,14 +39,14 @@ export GPUCI_CONDA_RETRY_SLEEP=30
 # SETUP - Check environment
 ################################################################################
 
-gpuci_logger "Get env"
+gpuci_logger "Check environment variables"
 env
 
 gpuci_logger "Activate conda env"
 . /opt/conda/etc/profile.d/conda.sh
 conda activate rapids
 
-gpuci_logger "Check versions"
+gpuci_logger "Check compiler versions"
 python --version
 $CC --version
 $CXX --version
@@ -56,28 +63,21 @@ conda config --set ssl_verify False
 # BUILD - Conda package builds (conda deps: libcupatial <- cuspatial)
 ##########################################################################################
 
-gpuci_logger "Build conda pkg for libcuspatial"
 if [ "$BUILD_LIBCUSPATIAL" == '1' ]; then
+  gpuci_logger "Build conda pkg for libcuspatial"
   if [[ -z "$PROJECT_FLASH" || "$PROJECT_FLASH" == "0" ]]; then
-    conda build conda/recipes/libcuspatial
+    gpuci_conda_retry build conda/recipes/libcuspatial
   else
-    [[ -e ${CUDF_HOME} ]] && rm -rf ${CUDF_HOME}
-    gpuci_logger "Clone cudf branch-$MINOR_VERSION"
-    git clone https://github.com/rapidsai/cudf.git -b branch-$MINOR_VERSION ${CUDF_HOME}
-    cd $CUDF_HOME
-    git submodule update --init --remote --recursive
-    cd $WORKSPACE
-
-    conda build --dirty --no-remove-work-dir conda/recipes/libcuspatial
+    gpuci_conda_retry build --dirty --no-remove-work-dir conda/recipes/libcuspatial
   fi
 fi
 
-gpuci_logger "Build conda pkg for cuspatial"
 if [ "$BUILD_CUSPATIAL" == '1' ]; then
+  gpuci_logger "Build conda pkg for cuspatial"
   if [[ -z "$PROJECT_FLASH" || "$PROJECT_FLASH" == "0" ]]; then
-    conda build conda/recipes/cuspatial
+    gpuci_conda_retry build conda/recipes/cuspatial
   else
-    conda build --dirty --no-remove-work-dir \
+    gpuci_conda_retry build --dirty --no-remove-work-dir \
         -c $WORKSPACE/ci/artifacts/cuspatial/cpu/conda-bld/ conda/recipes/cuspatial
   fi
 fi
@@ -86,6 +86,5 @@ fi
 # UPLOAD - Conda packages
 ################################################################################
 
-gpuci_logger "Upload packages"
+gpuci_logger "Upload conda pkgs..."
 source ci/cpu/upload.sh
-