+++ cpp/src/trajectory/trajectory_distances_and_speeds.cu
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-#include <thrust/iterator/discard_iterator.h>
+#include <cuspatial/error.hpp>
+#include <cuspatial/trajectory.hpp>
 
 #include <cudf/column/column_device_view.cuh>
 #include <cudf/column/column_factories.hpp>
@@ -25,8 +26,10 @@
 #include <cudf/utilities/traits.hpp>
 #include <cudf/utilities/type_dispatcher.hpp>
 
-#include <cuspatial/error.hpp>
-#include <cuspatial/trajectory.hpp>
+#include <rmm/cuda_stream_view.hpp>
+#include <rmm/exec_policy.hpp>
+
+#include <thrust/iterator/discard_iterator.h>
 
 namespace cuspatial {
 
@@ -42,7 +45,7 @@ struct duplicate_first_element_func {
 };
 
 template <typename T>
-auto duplicate_first_element_iterator(cudf::column_view const& col, cudaStream_t stream)
+auto duplicate_first_element_iterator(cudf::column_view const& col, rmm::cuda_stream_view stream)
 {
   auto d_col = cudf::column_device_view::create(col, stream);
   return thrust::make_transform_iterator(thrust::make_counting_iterator<cudf::size_type>(-1),
@@ -58,11 +61,9 @@ struct dispatch_timestamp {
     cudf::column_view const& x,
     cudf::column_view const& y,
     cudf::column_view const& timestamp,
-    rmm::mr::device_memory_resource* mr,
-    cudaStream_t stream)
+    rmm::cuda_stream_view stream,
+    rmm::mr::device_memory_resource* mr)
   {
-    auto policy = rmm::exec_policy(stream);
-
     // Construct output columns
     std::vector<std::unique_ptr<cudf::column>> cols{};
     cols.reserve(2);
@@ -81,7 +82,7 @@ struct dispatch_timestamp {
 
     using Rep     = typename Timestamp::rep;
     using Dur     = typename Timestamp::duration;
-    using Seconds = typename simt::std::chrono::seconds;
+    using Seconds = typename cuda::std::chrono::seconds;
 
     rmm::device_vector<Rep> durations(x.size() + 1);
     rmm::device_vector<double> distances(x.size() + 1);
@@ -102,7 +103,7 @@ struct dispatch_timestamp {
 
     // Compute duration and distance difference between adjacent elements that
     // share the same object id
-    thrust::adjacent_difference(policy->on(stream),                         // execution policy
+    thrust::adjacent_difference(rmm::exec_policy(stream),
                                 timestamp_point_and_id,                     // first
                                 timestamp_point_and_id + durations.size(),  // last
                                 duration_and_distance_1,                    // result
@@ -142,11 +143,11 @@ struct dispatch_timestamp {
     );
 
     using Period =
-      typename simt::std::ratio_divide<typename Timestamp::period, typename Seconds::period>::type;
+      typename cuda::std::ratio_divide<typename Timestamp::period, typename Seconds::period>::type;
 
     // Reduce the intermediate durations and kilometer distances into meter
     // distances and speeds in meters/second
-    thrust::reduce_by_key(policy->on(stream),               // execution policy
+    thrust::reduce_by_key(rmm::exec_policy(stream),
                           object_id.begin<int32_t>(),       // keys_first
                           object_id.end<int32_t>(),         // keys_last
                           duration_and_distance_2 + 1,      // values_first
@@ -165,7 +166,7 @@ struct dispatch_timestamp {
                           });
 
     // check for errors
-    CHECK_CUDA(stream);
+    CHECK_CUDA(stream.value());
 
     return std::make_unique<cudf::table>(std::move(cols));
   }
@@ -177,8 +178,8 @@ struct dispatch_timestamp {
     cudf::column_view const& timestamp,
     cudf::column_view const& x,
     cudf::column_view const& y,
-    rmm::mr::device_memory_resource* mr,
-    cudaStream_t stream)
+    rmm::cuda_stream_view stream,
+    rmm::mr::device_memory_resource* mr)
   {
     CUSPATIAL_FAIL("Timestamp must be a timestamp type");
   }
@@ -192,8 +193,8 @@ struct dispatch_element {
     cudf::column_view const& x,
     cudf::column_view const& y,
     cudf::column_view const& timestamp,
-    rmm::mr::device_memory_resource* mr,
-    cudaStream_t stream)
+    rmm::cuda_stream_view stream,
+    rmm::mr::device_memory_resource* mr)
   {
     return cudf::type_dispatcher(timestamp.type(),
                                  dispatch_timestamp<Element>{},
@@ -202,8 +203,8 @@ struct dispatch_element {
                                  x,
                                  y,
                                  timestamp,
-                                 mr,
-                                 stream);
+                                 stream,
+                                 mr);
   }
 
   template <typename Element>
@@ -213,8 +214,8 @@ struct dispatch_element {
              cudf::column_view const& x,
              cudf::column_view const& y,
              cudf::column_view const& timestamp,
-             rmm::mr::device_memory_resource* mr,
-             cudaStream_t stream)
+             rmm::cuda_stream_view stream,
+             rmm::mr::device_memory_resource* mr)
   {
     CUSPATIAL_FAIL("X and Y must be floating point types");
   }
@@ -228,11 +229,11 @@ std::unique_ptr<cudf::table> trajectory_distances_and_speeds(cudf::size_type num
                                                              cudf::column_view const& x,
                                                              cudf::column_view const& y,
                                                              cudf::column_view const& timestamp,
-                                                             rmm::mr::device_memory_resource* mr,
-                                                             cudaStream_t stream)
+                                                             rmm::cuda_stream_view stream,
+                                                             rmm::mr::device_memory_resource* mr)
 {
   return cudf::type_dispatcher(
-    x.type(), dispatch_element{}, num_trajectories, object_id, x, y, timestamp, mr, stream);
+    x.type(), dispatch_element{}, num_trajectories, object_id, x, y, timestamp, stream, mr);
 }
 }  // namespace detail
 
@@ -262,7 +263,7 @@ std::unique_ptr<cudf::table> trajectory_distances_and_speeds(cudf::size_type num
   }
 
   return detail::trajectory_distances_and_speeds(
-    num_trajectories, object_id, x, y, timestamp, mr, 0);
+    num_trajectories, object_id, x, y, timestamp, rmm::cuda_stream_default, mr);
 }
 
 }  // namespace cuspatial